# Conflicts:
#	marketing-cs/src/main/java/com/br/marketing/speedconfig/MarketingCommonConfig.java

+++ marketing-cs/src/main/java/com/br/marketing/mapper/MarketingTaskMapper.java
@@ -225,7 +225,7 @@ public interface MarketingTaskMapper extends MarketingTaskMapperBase {
     List<ScoreDetailVo> queryBatchs(CustomerBatchNumDTO dto);
 
     @AddDataAuth
-    Integer queryBatchsCount(CustomerBatchNumDTO dto);
+    Long queryBatchsCount(CustomerBatchNumDTO dto);
 
     MarketingTask selectCycleTopByApiCode(String apiCode);
